# server/docs/swagger.yaml
basePath: /
definitions:
...
  application.ApplicationSystem:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      name:
        description: 系统名
        type: string
      position:
        description: 系统位置
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  application.ApplicationSystemAdmin:
    properties:
      adminId:
        description: 管理员id
        type: integer
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      systemId:
        description: 系统id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  application.ApplicationSystemEditRelation:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      relation:
        description: 调用关系',
        type: string
      systemId:
        description: 系统id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  application.SystemRelation:
    properties:
      createdAt:
        description: 创建时间
        type: string
      endSystem:
        $ref: '#/definitions/application.ApplicationSystem'
      endSystemId:
        description: 目的节点id',
        type: integer
      endSystemUrl:
        description: 目的节点url',
        type: string
      id:
        description: 主键ID
        type: integer
      relation:
        description: 调用关系',
        type: string
      startSystem:
        $ref: '#/definitions/application.ApplicationSystem'
      startSystemId:
        description: 源节点id
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.AddSystem:
    properties:
      adminIds:
        items:
          type: integer
        type: array
      authorityId:
        description: 角色ID
        type: string
      system:
        $ref: '#/definitions/application.ApplicationSystem'
      systemAdmin:
        items:
          $ref: '#/definitions/application.ApplicationSystemAdmin'
        type: array
    type: object
...
  request.Empty:
    type: object
...
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: number
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      sortBy:
        type: string
      sortInverted:
        type: boolean
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.PageInfo:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /cmdb/addSystem:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统名, 位置, 管理员id, 主管
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AddSystem'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增系统
      tags:
      - CmdbSystem
...
  /cmdb/deleteSystem:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除系统
      tags:
      - CmdbSystem
  /cmdb/deleteSystemByIds:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除系统
      tags:
      - CmdbSystem
...
  /cmdb/getAdminSystems:
    post:
      consumes:
      - application/json
      parameters:
      - description: 空
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Empty'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取管理员所有系统
      tags:
      - CmdbSystem
...
  /cmdb/getSystemById:
    post:
      consumes:
      - application/json
      parameters:
      - description: 服务器id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取系统
      tags:
      - CmdbSystem
  /cmdb/getSystemList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取基础server列表
      tags:
      - CmdbSystem
...
  /cmdb/system/addEditRelation:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.ApplicationSystemEditRelation'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增编辑器关系图
      tags:
      - CmdbSystem
  /cmdb/system/addRelation:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.SystemRelation'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增联系
      tags:
      - CmdbSystem
  /cmdb/system/deleteEditRelation:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除编辑器关系图
      tags:
      - CmdbSystem
  /cmdb/system/getSystemEditRelations:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取编辑器关系图
      tags:
      - CmdbSystem
  /cmdb/system/relations:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 获取关系图
      tags:
      - CmdbSystem
  /cmdb/system/updateEditRelation:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/application.ApplicationSystemEditRelation'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新编辑器关系图
      tags:
      - CmdbSystem
...
  /cmdb/updateSystem:
    post:
      consumes:
      - application/json
      parameters:
      - description: 系统名, 位置, 管理员id, 主管
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AddSystem'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新服务器
      tags:
      - CmdbSystem
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"